user serviecs ts

+++ src/user/user.service.ts
@@ -1,19 +1,41 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { Repository } from 'typeorm';
 import { User } from './entities/user.entity';
 import { InjectRepository } from '@nestjs/typeorm';
+import { ConfigService } from '@nestjs/config';
+import * as bcrypt from 'bcrypt';
+import { envVariables } from 'src/common/const/env.const';
 
 @Injectable()
 export class UserService {
   constructor(
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
+    private readonly configService: ConfigService,
   ) {}
 
-  create(createUserDto: CreateUserDto) {
-    return this.userRepository.save(createUserDto);
+  async create(createUserDto: CreateUserDto) {
+    const { email, password } = createUserDto;
+    const user = await this.userRepository.findOne({ where: { email } });
+
+    if (user) {
+      throw new BadRequestException('User already exists');
+    }
+
+    const hash = await bcrypt.hash(
+      password,
+      this.configService.get<number>(envVariables.hashRounds),
+    );
+
+    await this.userRepository.save({ email, password: hash });
+
+    return this.userRepository.findOne({ where: { email } });
   }
 
   findAll() {
@@ -24,8 +46,10 @@ export class UserService {
     const user = await this.userRepository.findOne({ where: { id } });
 
     if (!user) {
-      return new NotFoundException('User not found');
+      throw new NotFoundException('User not found');
     }
+
+    return user;
 
     // return this.userRepository.findOne({ where: { id } });
   }
@@ -34,21 +58,31 @@ export class UserService {
     const user = await this.userRepository.findOne({ where: { id } });
 
     if (!user) {
-      return new NotFoundException('User not found');
+      throw new NotFoundException('User not found');
     }
 
-    await this.userRepository.update(id, updateUserDto);
+    const hash = await bcrypt.hash(
+      updateUserDto.password,
+      this.configService.get<number>(envVariables.hashRounds),
+    );
+
+    await this.userRepository.update(id, {
+      ...updateUserDto,
+      password: hash,
+    });
 
     return this.userRepository.findOne({ where: { id } });
   }
 
   async remove(id: number) {
-    // const user = await this.userRepository.findOne({ where: { id } });
-    //
-    // if (!user) {
-    //   return new NotFoundException('User not found');
-    // }
+    const user = await this.userRepository.findOne({ where: { id } });
 
-    return this.userRepository.delete(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    await this.userRepository.delete(id);
+
+    return id;
   }
 }
